Minor changes in datamodelSlice.ts to make it match the other slices
