+++ src/altinn-app-frontend/src/features/form/datamodel/index.d.ts
@@ -1,0 +1,17 @@
+export interface IJsonSchemas {
+  [id: string]: object;
+}
+
+export interface IDataModelState {
+  schemas: IJsonSchemas;
+  error: Error;
+}
+
+export interface IFetchJsonSchemaFulfilled {
+  schema: object;
+  id: string;
+}
+
+export interface IFetchJsonSchemaRejected {
+  error: Error;
+}
