+++ src/altinn-app-frontend/src/shared/resources/queue/stateless/statelessQueueSaga.tsx
@@ -1,7 +1,7 @@
 import type { SagaIterator } from '@redux-saga/types';
 import { put, takeLatest } from 'redux-saga/effects';
 import { FormDataActions } from 'src/features/form/data/formDataSlice';
-import { fetchJsonSchema } from 'src/features/form/datamodel/datamodelSlice';
+import { DataModelActions } from 'src/features/form/datamodel/datamodelSlice';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
 import { startStatelessIsLoading } from '../../isLoading/isLoadingSlice';
 import {
@@ -14,7 +14,7 @@ export function* startInitialStatelessQueueSaga(): SagaIterator {
   try {
     yield put(startStatelessIsLoading());
     yield put(FormDataActions.fetchInitial());
-    yield put(fetchJsonSchema());
+    yield put(DataModelActions.fetchJsonSchema());
     yield put(FormLayoutActions.fetchSets());
     yield put(FormLayoutActions.fetch());
     yield put(FormLayoutActions.fetchSettings());
